fix: Stage3 재구성 및 re-ordering 완료
- stage3 재구성
- 난이도에 따라 stage3 <-> stage2 스왑완료
- 스테이지 스왑에 따른 스테이지 가이드 에셋파일 업데이트
