+++ src/data/stageThreeCoordinates.json
@@ -1,110 +1,104 @@
 {
-  "departure": [0, 0.5, 0],
-  "arrival": [8, 0.5, 0],
+  "departure": [0, 0.5, 4],
+  "arrival": [4, 0.5, 8],
+
   "cubes": {
     "positions": [
-      { "id": "stageThree-cube-1", "coordinate": [0, 0.5, 0] },
-      { "id": "stageThree-cube-2", "coordinate": [0, 0.5, 1] },
-      { "id": "stageThree-cube-3", "coordinate": [0, 0.5, 2] },
-      { "id": "stageThree-cube-4", "coordinate": [1, 0.5, 0] },
-      { "id": "stageThree-cube-5", "coordinate": [1, 0.5, 1] },
-      { "id": "stageThree-cube-6", "coordinate": [1, 0.5, 2] },
-      { "id": "stageThree-cube-7", "coordinate": [2, 0.5, 0] },
-      { "id": "stageThree-cube-8", "coordinate": [2, 0.5, 1] },
-      { "id": "stageThree-cube-9", "coordinate": [2, 0.5, 2] },
-      { "id": "stageThree-cube-10", "coordinate": [4, 0.5, 0] },
-      { "id": "stageThree-cube-11", "coordinate": [4, 0.5, 1] },
-      { "id": "stageThree-cube-12", "coordinate": [4, 0.5, 2] },
-      { "id": "stageThree-cube-13", "coordinate": [5, 0.5, 0] },
-      { "id": "stageThree-cube-14", "coordinate": [5, 0.5, 1] },
-      { "id": "stageThree-cube-15", "coordinate": [5, 0.5, 2] },
-      { "id": "stageThree-cube-16", "coordinate": [6, 0.5, 0] },
-      { "id": "stageThree-cube-17", "coordinate": [6, 0.5, 1] },
-      { "id": "stageThree-cube-18", "coordinate": [6, 0.5, 2] },
-      { "id": "stageThree-cube-19", "coordinate": [8, 0.5, 0] },
-      { "id": "stageThree-cube-20", "coordinate": [8, 0.5, 1] },
-      { "id": "stageThree-cube-21", "coordinate": [8, 0.5, 2] },
-      { "id": "stageThree-cube-22", "coordinate": [1, 1.5, 1] },
-      { "id": "stageThree-cube-23", "coordinate": [1, 1.5, 0] },
-      { "id": "stageThree-cube-24", "coordinate": [5, 1.5, 1] },
-      { "id": "stageThree-cube-25", "coordinate": [5, 1.5, 2] }
-    ]
-  },
-  "hiddenCubes": {
-    "positions": [
-      { "id": "stageThree-invisibleCube-1", "coordinate": [6, 5.5, 6] },
-      { "id": "stageThree-invisibleCube-2", "coordinate": [6, 5.5, 7] },
-      { "id": "stageThree-invisibleCube-3", "coordinate": [6, 5.5, 8] },
-      { "id": "stageThree-invisibleCube-4", "coordinate": [7, 5.5, 6] },
-      { "id": "stageThree-invisibleCube-5", "coordinate": [7, 5.5, 7] },
-      { "id": "stageThree-invisibleCube-6", "coordinate": [7, 5.5, 8] },
-      { "id": "stageThree-invisibleCube-7", "coordinate": [8, 5.5, 6] },
-      { "id": "stageThree-invisibleCube-8", "coordinate": [8, 5.5, 7] },
-      { "id": "stageThree-invisibleCube-9", "coordinate": [8, 5.5, 8] },
-      { "id": "stageThree-invisibleCube-10", "coordinate": [10, 5.5, 6] },
-      { "id": "stageThree-invisibleCube-11", "coordinate": [10, 5.5, 7] },
-      { "id": "stageThree-invisibleCube-12", "coordinate": [10, 5.5, 8] },
-      { "id": "stageThree-invisibleCube-13", "coordinate": [11, 5.5, 6] },
-      { "id": "stageThree-invisibleCube-14", "coordinate": [11, 5.5, 7] },
-      { "id": "stageThree-invisibleCube-15", "coordinate": [11, 5.5, 8] },
-      { "id": "stageThree-invisibleCube-16", "coordinate": [12, 5.5, 6] },
-      { "id": "stageThree-invisibleCube-17", "coordinate": [12, 5.5, 7] },
-      { "id": "stageThree-invisibleCube-18", "coordinate": [12, 5.5, 8] },
-      { "id": "stageThree-invisibleCube-19", "coordinate": [7, 6.5, 6] },
-      { "id": "stageThree-invisibleCube-20", "coordinate": [7, 6.5, 7] },
-      { "id": "stageThree-invisibleCube-21", "coordinate": [11, 6.5, 6] },
-      { "id": "stageThree-invisibleCube-22", "coordinate": [11, 6.5, 7] }
+      { "id": "stageTwo-cube-1", "coordinate": [4, 0.5, 0] },
+      { "id": "stageTwo-cube-2", "coordinate": [8, 0.5, 4] },
+      { "id": "stageTwo-cube-3", "coordinate": [4, 0.5, 8] },
+      { "id": "stageTwo-cube-4", "coordinate": [0, 0.5, 4] },
+      { "id": "stageTwo-cube-5", "coordinate": [3, 3.5, 3] },
+      { "id": "stageTwo-cube-6", "coordinate": [4, 3.5, 3] },
+      { "id": "stageTwo-cube-7", "coordinate": [5, 3.5, 3] },
+      { "id": "stageTwo-cube-8", "coordinate": [3, 3.5, 4] },
+      { "id": "stageTwo-cube-9", "coordinate": [4, 3.5, 4] },
+      { "id": "stageTwo-cube-10", "coordinate": [5, 3.5, 4] },
+      { "id": "stageTwo-cube-11", "coordinate": [3, 3.5, 5] },
+      { "id": "stageTwo-cube-12", "coordinate": [4, 3.5, 5] },
+      { "id": "stageTwo-cube-13", "coordinate": [5, 3.5, 5] },
+      { "id": "stageTwo-cube-14", "coordinate": [3, 4.5, 4] },
+      { "id": "stageTwo-cube-15", "coordinate": [4, 4.5, 3] },
+      { "id": "stageTwo-cube-16", "coordinate": [4, 4.5, 4] },
+      { "id": "stageTwo-cube-17", "coordinate": [4, 4.5, 5] },
+      { "id": "stageTwo-cube-18", "coordinate": [5, 4.5, 4] }
     ]
   },
   "linkEdges": [
     {
-      "id": "stageThree-linkEdge-1",
+      "id": "stageTwo-linkEdge-1",
       "edgeFrom": {
-        "pointA": [5.5, 6, 5.5],
-        "pointB": [6.5, 6, 5.5]
+        "pointA": [2.5, 4, 2.5],
+        "pointB": [2.5, 4, 3.5]
       },
       "edgeTo": {
-        "pointA": [1.5, 1, 2.5],
-        "pointB": [2.5, 1, 2.5]
+        "pointA": [0.5, 1, 3.5],
+        "pointB": [0.5, 1, 4.5]
       },
       "color": "blue",
       "thickness": 3
     },
     {
-      "id": "stageThree-linkEdge-2",
+      "id": "stageTwo-linkEdge-2",
       "edgeFrom": {
-        "pointA": [7.5, 6, 5.5],
-        "pointB": [8.5, 6, 5.5]
+        "pointA": [2.5, 4, 2.5],
+        "pointB": [3.5, 4, 2.5]
       },
       "edgeTo": {
-        "pointA": [3.5, 1, 2.5],
-        "pointB": [4.5, 1, 2.5]
+        "pointA": [3.5, 1, 0.5],
+        "pointB": [4.5, 1, 0.5]
       },
       "color": "red",
       "thickness": 3
     },
     {
-      "id": "stageThree-linkEdge-3",
+      "id": "stageTwo-linkEdge-3",
       "edgeFrom": {
-        "pointA": [9.5, 6, 5.5],
-        "pointB": [10.5, 6, 5.5]
+        "pointA": [4.5, 4, 2.5],
+        "pointB": [5.5, 4, 2.5]
       },
       "edgeTo": {
-        "pointA": [5.5, 1, 2.5],
-        "pointB": [6.5, 1, 2.5]
+        "pointA": [3.5, 1, 0.5],
+        "pointB": [4.5, 1, 0.5]
+      },
+      "color": "yellow",
+      "thickness": 3
+    },
+    {
+      "id": "stageTwo-linkEdge-4",
+      "edgeFrom": {
+        "pointA": [5.5, 4, 2.5],
+        "pointB": [5.5, 4, 3.5]
+      },
+      "edgeTo": {
+        "pointA": [7.5, 1, 3.5],
+        "pointB": [7.5, 1, 4.5]
+      },
+      "color": "purple",
+      "thickness": 3
+    },
+    {
+      "id": "stageTwo-linkEdge-5",
+      "edgeFrom": {
+        "pointA": [4.5, 4, 5.5],
+        "pointB": [5.5, 4, 5.5]
+      },
+      "edgeTo": {
+        "pointA": [3.5, 1, 7.5],
+        "pointB": [4.5, 1, 7.5]
       },
       "color": "green",
       "thickness": 3
     },
     {
-      "id": "stageThree-linkEdge-4",
+      "id": "stageTwo-linkEdge-6",
       "edgeFrom": {
-        "pointA": [11.5, 6, 5.5],
-        "pointB": [12.5, 6, 5.5]
+        "pointA": [5.5, 4, 4.5],
+        "pointB": [5.5, 4, 5.5]
       },
       "edgeTo": {
-        "pointA": [7.5, 1, 2.5],
-        "pointB": [8.5, 1, 2.5]
+        "pointA": [7.5, 1, 3.5],
+        "pointB": [7.5, 1, 4.5]
       },
       "color": "black",
       "thickness": 3
